Fixes #409 Remove scopes
The goal is to simplify the package for the v6. Scopes are useless: I don't see a valid use case for the "prototype" scope, if needed then it should be replaced by an actual factory.

By removing scopes the API would be simpler as well as the internals of PHP-DI.